Move product related endpoints to seperate service part
